backend/backend_requirements.txt
django==4.2.3
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysqlclient==2.2.7
requests==2.32.3
python-dotenv==1.0.0
gunicorn==21.2.0
docker==6.1.3
psutil==5.9.5
Pillow==11.1.0
jupyter==1.0.0
pandas==2.1.0
numpy==1.25.2
openpyxl==3.1.2
